+++ app/src/main/java/lavender/client/android/ui/chatlist/ChatListActivity.kt
@@ -120,6 +120,7 @@ class ChatListActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+        androidx.core.view.WindowCompat.setDecorFitsSystemWindows(window, false)
 
         SessionManager.initFromPrefs(this)
         applyTheme()